% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{anti_t}
\alias{anti_t}
\title{Compute the anti-transpose of a square matrix or table}
\usage{
anti_t(x)
}
\arguments{
\item{x}{A square matrix or table.}
}
\value{
The anti-transpose (transpose across the off-diagonal) of \code{x}.
}
\description{
The anti-transpose of a square matrix or table \code{x} is the
transpose taken across the off-diagonal.  \code{anti-t} is a convenience
wrapper for \code{apply(apply(x, 2, rev), 1, rev)}.

The anti-transpose function is useful, for example, for converting the
results of \code{table(Test, Gold)} to a "confusion" matrix, and vice-versa.

A C implementation of the anti-transpose function can also be found as in
package \code{gclm} as \code{gclm:::anti_t}.
}
\examples{
Gold <- rbinom(20, 1, 0.50)
Test <- Gold; Test[c(3, 5, 9, 12, 16)] <- 1 - Test[c(3, 9, 5, 12, 16)]
(tb <- table(Test, Gold))
anti_t(tb)
}
