% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{BBS_mortality}
\alias{BBS_mortality}
\title{Estimate Probability of Mortality Among Key Population Members}
\usage{
BBS_mortality(
  .data,
  n_know = NULL,
  n_lost = NULL,
  n_died = NULL,
  degree = NULL,
  recall_max = 100,
  R = 2000L,
  conf = 0.95,
  ci_type = c("bca", "perc", "basic")
)
}
\arguments{
\item{.data}{a dataframe containing the four variables.}

\item{n_know}{a character string naming the variable containing the number
of contacts known by the subject and who know the subject within the past
time interval.}

\item{n_lost}{a character string naming the variable containing the subset
of \code{n_know} for whom the subject lost contact during the past time
interval.}

\item{n_died}{a character string naming the variable containing the subset
of \code{n_know} who were known by the subject to have died, by any cause,
during the past time interval.}

\item{degree}{a character string naming the variable containing the network
degree of the subject.}

\item{recall_max}{the maximum contact count that is deemed plausible.
Default: 100.}

\item{R}{an integer-valued count of the number of bootstrap replicates
desired for estimation of confidence limits. Default: 2000.}

\item{conf}{the desired confidence level. Default: 0.95.}

\item{ci_type}{a character string naming the type of bootstrap confidence
interval desired; one of \verb{'bca'} (default), \verb{'perc'} or
\verb{'basic'}.  See \code{\link[boot]{boot.ci}}.}
}
\value{
A dataframe containing the following variables:
\describe{
\item{\code{estimate}}{The estimated probability of mortality}
\item{\code{SE}}{The standard error of the estimate}
\item{\code{lower}}{The lower confidence limit}
\item{\code{upper}}{The upper confidence limit}
\item{\code{conf_level}}{The confidence level}
\item{\code{type}}{The type of bootstrap confidence interval}
}
}
\description{
\code{BBS_mortality} computes estimates of the probability of
death over some time interval (determined by the survey questions) among
members of key populations using information about contacts collected as part
of Bio-Behavioral Surveillance Surveys. Estimation includes bootstrap
confidence intervals.
}
\details{
\code{BBS_mortality} eliminates records for which
\code{n_know} < \code{n_lost} + \code{n_died} and any observations containing
missing values for those variables are also removed.
}
\author{
Ian Fellows and Steve Gutreuter
}
