% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{binom_coef}
\alias{binom_coef}
\title{Compute the binomial coefficient \eqn{\binom{n}{k}}}
\usage{
binom_coef(n, k)
}
\arguments{
\item{n}{a non-negative number}

\item{k}{another non-negative number such that \eqn{k \le n}}
}
\value{
Numeric value of the binomial coefficient \eqn{\binom{n}{k} = \frac{n!}{k!(n - k)!}}

#'
}
\description{
Compute the binomial coefficient \eqn{\binom{n}{k}}
}
\author{
Steve Gutreuter
}
