% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{binom_expansion}
\alias{binom_expansion}
\title{Compute the binomial expansion \eqn{(x + y)^n}}
\usage{
binom_expansion(x, y, n)
}
\arguments{
\item{x}{A real number}

\item{y}{A real number}

\item{n}{integer-valued exponent}
}
\value{

}
\description{
Compute the binomial expansion \eqn{(x + y)^n}
}
\author{
Steve Gutreuter
}
