% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{computeBetaMoments}
\alias{computeBetaMoments}
\title{Compute the mean and variance of the Beta distribution from the shape
parameters}
\usage{
computeBetaMoments(a, b)
}
\arguments{
\item{a}{Beta distribution shape1 parameter}

\item{b}{Beta distribution shape2 parameter}
}
\value{
A list containing the mean and variance
}
\description{
Compute the mean and variance of the Beta distribution from the shape
parameters
}
\author{
Steve Gutreuter
}
