% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{computeBetaParms}
\alias{computeBetaParms}
\title{Compute the shape parameters of the Beta distribution from the mean and
variance}
\usage{
computeBetaParms(mean, var)
}
\arguments{
\item{mean}{Mean of the Beta distribution}

\item{var}{Variance of the Beta distribution}
}
\value{
A list containing the two shape parameters of the Beta distribution
}
\description{
Compute the shape parameters of the Beta distribution from the mean and
variance
}
\author{
Steve Gutreuter
}
