% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{computeGammaMoments}
\alias{computeGammaMoments}
\title{Compute the mean and variance of the Gamma distribution from the
shape and rate or scale parameters}
\usage{
computeGammaMoments(shape, rate = 1, scale = 1/rate)
}
\arguments{
\item{shape}{Gamma distribution shape parameter}

\item{rate}{Gamma distribution rate parameter}

\item{scale}{Gamma distribution scale = 1 / rate}
}
\value{
A list containing the mean and variance of the Gamma
distribution
}
\description{
Compute the mean and variance of the Gamma distribution from the
shape and rate or scale parameters
}
\author{
Steve Gutreuter
}
