% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{computeGammaParms}
\alias{computeGammaParms}
\title{Compute the mean and variance of the Gamma distribution from the
shape and rate or scale parameters}
\usage{
computeGammaParms(mu, var)
}
\arguments{
\item{mu}{Gamma distribution mean}

\item{var}{Gamma distribution variance}
}
\value{
A list containing the shape, rate and scale parameters of
Gamma distribution
}
\description{
Compute the mean and variance of the Gamma distribution from the
shape and rate or scale parameters
}
\author{
Steve Gutreuter
}
