% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{count_NA}
\alias{count_NA}
\title{Count the numbers of NA, NaN and Inf in each column of a dataframe}
\usage{
count_NA(.data)
}
\arguments{
\item{.data}{A dataframe or tibble}
}
\value{
A matrix containing counts of NA, NaN and Inf values.
}
\description{
Count the numbers of NA, NaN and Inf in each column of a dataframe
}
\author{
Steve Gutreuter
}
\keyword{EDA,}
\keyword{NA}
\keyword{missing}
\keyword{values,}
