% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{empCDF}
\alias{empCDF}
\title{Compute the discrete empirical cumulative distribution of a real vector}
\usage{
empCDF(x, knots = 10)
}
\arguments{
\item{x}{A numeric vector}

\item{knots}{Number of breaks in the abscissa}
}
\value{
A numeric vector
}
\description{
Compute the discrete empirical cumulative distribution of a real vector
}
\author{
Steve Gutreuter
}
