% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{fp_table}
\alias{fp_table}
\title{One and two-way summaries of frequencies, percentages and totals.}
\usage{
fp_table(
  .data,
  colvar,
  rowvar = NULL,
  transpose = FALSE,
  useNA = c("no", "ifany", "always")
)
}
\arguments{
\item{.data}{A dataframe or tibble.}

\item{colvar}{Character string naming the factor variable for tabulation.}

\item{rowvar}{(Optional) string naming factor the row variable for cross-tabulation.}

\item{transpose}{Logical specifying whether to transpose the result.}

\item{useNA}{One of \verb{"no"} (default), \verb{"ifany"} or \verb{"always"} specifying
the handling of missing values.}
}
\value{
A dataframe containing frequencies and percentages. The columns
contain frequecies and row percentages.  A two-way summary is produced if
rowvar is specified.  Rows and columns can be reversed by specifying
\code{transpose = TRUE}. The result is uglier than a hairless cat, but can
easily be converted to useful tabular formats with some custom recoding.
}
\description{
One and two-way summaries of frequencies, percentages and totals.
}
\author{
Steve Gutreuter
}
\keyword{cross-tabulation}
\keyword{frequencies}
\keyword{percentages}
\keyword{table}
