% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{icc2deff}
\alias{icc2deff}
\title{Compute the conventional design effect (Deff) from the intraclass correlation}
\usage{
icc2deff(icc, N)
}
\arguments{
\item{icc}{Numeric intraclass (intracluster) correlation coefficient}

\item{N}{The average number of elements per class/cluster}
}
\value{
The design effect (Deff)
}
\description{
Compute the conventional design effect (Deff) from the intraclass correlation
}
\author{
Steve Gutreuter
}
