% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFunctions.R
\name{mid_date}
\alias{mid_date}
\title{Compute the midpoint between two dates}
\usage{
mid_date(startdate, enddate)
}
\arguments{
\item{startdate}{A vector of class \code{Date} containing the beginning
date(s)}

\item{enddate}{A vector of class \code{Date} containing the ending date(s)}
}
\value{
A vector of class \code{Date} containing the midpoint date(s) between
\code{begindate} and \code{enddate}
}
\description{
Compute the midpoint between two dates
}
\examples{
begin <- ymd("2020-06-21", "2021-06-21")
end <- ymd("2020-12-21", "2021-12-21")
mid_date(begin, end)

}
\seealso{
\code{\link[lubridate]{interval}}
}
\author{
Steve Gutreuter
}
