% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionFunctions.R
\name{nuke_packages}
\alias{nuke_packages}
\title{Detach all packages other than those attached by base R}
\usage{
nuke_packages()
}
\description{
This function is used for its side effects and returns nothing.  Do no harm
to others; call this function only within your own interactive session.
Some namespaces may not be unloaded.
}
