% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{Oz_incidencer}
\alias{Oz_incidencer}
\title{Estimate HIV incidence rate using "Osmond's" method}
\usage{
Oz_incidencer(
  DF,
  x = NULL,
  left = NULL,
  right = NULL,
  surveyDate = NULL,
  debutDate = NULL,
  sweight = NULL,
  imputer = "random",
  reps = 1001L,
  ci.type = "bca",
  conf = 0.95,
  ...
)
}
\arguments{
\item{DF}{A dataframe containing the disease indicator and an
estimate of time at risk.}

\item{x}{Character-valued name of the binary (0 = negative, 1 = positive)
infection status indicator.}

\item{left}{Character-valued name of the (class = \verb{"Date"}) variable for
the earliest possble date of infection.}

\item{right}{Character-valued name of the (class = \verb{"Date"}) variable for
the latest possible date of infection.}

\item{surveyDate}{Character-valued name of the (class = \verb{"Date"}) variable
containing the interview dates.}

\item{debutDate}{Character-valued name of the (class = \verb{"Date"}) variable
containing the self-reported dates of sexual debut.}

\item{sweight}{(Optional) Numeric survey sampling weights.}

\item{imputer}{Character; either \verb{"random"} (the default) or
\verb{"midpoint"} which specifies how the date of infection is imputed.  With
argument \code{imputer = "random"}, an independently and identically
uniformly distributed date of infection is imputed for each
bootstrap replication.  This is slow for reasonable numbers of bootstrap
replications.}

\item{reps}{Integer number of bootstrap replicates.}

\item{ci.type}{Bootstrap confidence interval type; any subset of
\code{c{"norm","basic", "stud", "perc", "bca"}}, or \code{"all"}.  The default
is \verb{"bca"}, which may require large numbers of bootstrap replicates.}

\item{conf}{Numeric confidence level for bootstrap intervals.}

\item{...}{Additional arguments passed to \code{boot} or \code{boot.ci}}
}
\value{
A list containing:
\describe{
\item{\verb{boot.res}}{An object of class \verb{boot} including the incidence
rate estimate.}
\item{\verb{boot.ci}}{An object of class \verb{bootci} containng the
 bootstrap confidence intervals.}
\item{\verb{Call}}{The function call.}
}
}
\description{
Estimate HIV incidence rate (infections per person-year) as the simple ratio
of the numbers of infections to the total imputed time at risk.  Frequentist
confidence intervals are computed by bootstrapping.  Dates of infection are
known only within intervals, and individual times-at-risk are imputed as the
disease-free periods between sexual debut and either as uniformly distributed
random dates between dates of sexual debut and diagnosis or as the midpoints
between those dates.
}
\references{
Osmond DH, Page K, Wiley J, Garrett K, Sheppard HW, Moss AR, Schrager L,
Winkelstein W.  HIV infection in homosexual and bisexual men 18 to 29
years of age: The San Francisco Young Men's Health Study.  American
Journal of Public Health 1994;84(12):1933-1937.

Davison AC, Hinkley DV.  Bootstrap Methods and Their Application, Chapter 5.
Cambridge University Press; 1997.
}
\seealso{
\code{\link{boot}}
}
\author{
Steve Gutreuter, \email{sgutreuter@cdc.com}
}
