% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFunctions.R
\name{rand_date}
\alias{rand_date}
\title{Compute uniformly distributed random date(s) between two dates}
\usage{
rand_date(startdate, enddate, size = 1, simplify = FALSE)
}
\arguments{
\item{startdate}{A vector of class \code{Date} containing the beginning
date(s)}

\item{enddate}{A vector of class \code{Date} containing the ending date(s)}

\item{size}{An integer vector of the number of random dates}

\item{simplify}{(Logical) controlling whether the resulting list is
simplfied to a vector or matrix}
}
\value{
A list or vector of elements of class \code{Date} containing uniformly
distributed dates in the closed interval (\code{begindate}, \code{enddate}).
}
\description{
Compute uniformly distributed random date(s) between two dates
}
\examples{
begin <- ymd("2020-06-21", "2021-06-21")
end <- ymd("2020-12-21", "2021-12-21")
rand_date(begin, end, size = 5)

rand_date(begin, end, simplify = TRUE)

}
\seealso{
\code{link(lubridate::interval)}
}
\author{
Steve Gutreuter
}
