% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{recode_if}
\alias{recode_if}
\title{Recode a variable conditional on one or more other variables}
\usage{
recode_if(.x, .condition, ...)
}
\arguments{
\item{.x}{A vector to modify}

\item{.condition}{An expression of a condition}

\item{...}{Replacement (see dplyr::recode)}
}
\value{
Contents of "Value"
}
\description{
Recode a variable conditional on one or more other variables
}
\references{
\url{https://gerkelab.com/blog/2018/08/recode_if}
}
\seealso{
\code{\link{dplyr}}
}
\author{
Unknown
}
\keyword{recode}
