% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{sampsize_DHS}
\alias{sampsize_DHS}
\title{Compute the stratum-specific sample size for a proportion estimated from a
Demographic and Health Survey}
\usage{
sampsize_DHS(p = 0.5, Deff = 2, RSE = 1, R1 = 1, R2 = 1, eligibles = 1)
}
\arguments{
\item{p}{A vector of anticipated proportions}

\item{Deff}{The anticipated design effect}

\item{RSE}{The desired relative standard error of estimation}

\item{R1}{Household response fraction}

\item{R2}{Individual response fraction}

\item{eligibles}{The anticipated average number of eligible respondents per
household}
}
\value{
A numeric vector of estimated sample sizes
}
\description{
Compute the stratum-specific sample size for a proportion estimated from a
Demographic and Health Survey
}
\seealso{
\url{https://dhsprogram.com/publications/publication-dhsg1-dhs-questionnaires-and-manuals.cfm}
}
\author{
Steve Gutreuter
}
