% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{sampsize_multinomial}
\alias{sampsize_multinomial}
\title{Compute the approximate worst-case sample size for a vector of multinomial
proportions}
\usage{
sampsize_multinomial(m, relmoe, conf = 0.95)
}
\arguments{
\item{m}{The (integer) number of elements (at least two) in the vector of
equal proportions.}

\item{relmoe}{Relative margin of error (relative half-width of the confidence
interval), expressed as a proportion.  A value of 0.25 specifies margin of
error equal to 0.25 times the value of the (equal) proportions.}

\item{conf}{Confidence level.  The default is 0.95.}
}
\value{
A list containing the following elements:
\describe{
\item{\verb{sample_size}}{The required sample size.}
\item{\verb{target_proportions}}{The anticipated vector of multinomially
distributed outcomes.}
\item{\verb{moe}}{The absolute margins of error for estimation of
\verb{target_proportions}.}
\item{\verb{Call}}{The function call.}
}
}
\description{
Compute the approximate worst-case sample size for a vector of multinomial
proportions
}
\references{
Thompson SK. Sample size for estimating multinomial proportions. The American
Statistician 1987; 41(1):42-46.
}
\author{
Steve Gutreuter
}
