% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGmisc-package.R
\docType{package}
\name{SGmisc}
\alias{SGmisc}
\title{SGmisc}
\description{
A collection of small wrappers and convienience functions.
}
\seealso{
The exported functions in the \code{SGmisc} package are:
\describe{
\item{\code{\link{computeBetaMoments}}}{Compute the mean oand variance of the
Beta distribution}
\item{\code{\link{computeBetaParms}}}{Compute the shape parameters of the
Beta distribution}
\item{\code{\link{count_NA}}}{Get a count of the numbers of NA, NaN and Inf
for each column of a dataframe}
\item{\code{\link{empCDF}}}{Compute the empirical cumulative distribution
function for a discrete random vector}
\item{\code{\link{fp_table}}}{Create one- and two-way tables of frequencies and
percentages}
\item{\code{\link{icc2deff}}}{Compute the conventional survey design effect
from the intraclass correlation}
\item{\code{\link{ilogit}}}{Compute the inverse of the logit transformation
("expit" function)}
\item{\code{\link{logit}}}{Compute the logit transformation of a vector of
elements in (0, 1)}
\item{\code{\link{mid_date}}}{Find the midpoint date between two dates}
\item{\code{\link{Oz_incidencer}}}{Compute HIV incidence rate using
"Osmond's" method}
\item{\code{\link{rand_date}}}{Compute a random uniformly distributed date
between two dates}
\item{\code{\link{recode_if}}}{Conditional recoding of elements of a vector}
\item{\code{\link{sampsize_DHS}}}{Compute a sample-size requirement from for
a Demographic and Health Survey}
\item{\code{\link{sampsize_multinomial}}}{Compute the approximate worst-case
sample-size for a vector of multinomially distributed proportions}
\item{\code{\link{smooth_extremum}}}{Compute a smooth differentiable approximation
to the minimum or maximum of a numeric vector}
}
}
\author{
Steve Gutreuter: \email{sgutreuter@gmail.com}
}
