% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{smooth_extremum}
\alias{smooth_extremum}
\title{Compute a smooth differentiable approximation to the minimum or maximum of a
numeric vector}
\usage{
smooth_extremum(x, type = NULL, alpha = 4)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{Character \code{"min"} or \code{"max"} specifying the type of
extremum desired.}

\item{alpha}{A numeric smoothing parameter (\code{alpha} > 0).}
}
\value{
An approximation to the minimum or maximum of \code{x}.
}
\description{
Compute a smooth differentiable approximation to the minimum or maximum of a
numeric vector
}
\references{
\url{https://en.wikipedia.org/wiki/Smooth_maximum}
}
