% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFunctions.R
\name{timer}
\alias{timer}
\title{Report starting, ending and elapsed times of execution of code}
\usage{
timer(start = t0_)
}
\arguments{
\item{start}{A datetime value containing the starting time (Default
t0_)}

\item{enddate}{A vector of class \code{Date} containing the ending date(s)}
}
\value{
Print to console the starting, ending and elapsed times
}
\description{
Report starting, ending and elapsed times of execution of code
}
\examples{
t0_ <- Sys.time()
Sys.sleep(10)
timer()
}
\author{
Steve Gutreuter
}
