% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{ilogit}
\alias{ilogit}
\alias{expit}
\title{Inverse logit (expit) tranformation of a numeric vector}
\usage{
ilogit(x)

expit(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector of the inverse logits of x.
}
\description{
Inverse logit (expit) tranformation of a numeric vector
}
\author{
Steve Gutreuter
}
\keyword{expit}
