% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericFunctions.R
\name{logit}
\alias{logit}
\title{Logit transformation of a real vector}
\usage{
logit(x)
}
\arguments{
\item{x}{A real-valued vector with elements in (0,1)}
}
\value{
A numeric vector containing logit(x)
}
\description{
Logit transformation of a real vector
}
\author{
Steve Gutreuter
}
