% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFunctions.R
\name{rand_date}
\alias{rand_date}
\title{Compute uniformly distributed random dates within intervals of dates}
\usage{
rand_date(startdate, enddate, ...)
}
\arguments{
\item{startdate}{A vector of class \code{Date} containing the beginning
date(s)}

\item{enddate}{A vector of class \code{Date} containing the ending date(s)}

\item{...}{Additional arguments passed to \code{base::sample}}
}
\value{
A  vector of elements of class \code{Date} containing uniformly
distributed dates in the closed interval (\code{startdate}, \code{enddate}).
}
\description{
Compute uniformly distributed random dates within intervals of dates
}
\examples{
begin <- ymd("2020-06-21", "2021-06-21")
end <- ymd("2020-12-21", "2021-12-21")
rand_date(begin, end)

}
\seealso{
\code{\link[base]{sample}} \code{\link[base]{seq.Date}}
}
\author{
Steve Gutreuter
}
